decouple contribution fields from exact target payloads
