+++ handlers/new-product-api/src/main/scala/com/gu/newproduct/api/addsubscription/email/SendConfirmationEmail.scala
@@ -6,7 +6,7 @@ import com.gu.effects.sqs.AwsSQSSend.Payload
 import com.gu.i18n.Currency
 import com.gu.newproduct.api.addsubscription.TypeConvert._
 import com.gu.newproduct.api.addsubscription.ZuoraAccountId
-import com.gu.newproduct.api.addsubscription.zuora.GetContacts.{Contacts, Email}
+import com.gu.newproduct.api.addsubscription.zuora.GetContacts.Contacts
 import com.gu.newproduct.api.addsubscription.zuora.GetPaymentMethod.DirectDebit
 import com.gu.util.Logging
 import com.gu.util.apigateway.ApiGatewayResponse
@@ -14,6 +14,7 @@ import com.gu.util.reader.AsyncTypes._
 import com.gu.util.reader.Types.ApiGatewayOp.{ContinueProcessing, ReturnWithResponse}
 import com.gu.util.resthttp.Types.ClientFailableOp
 import play.api.libs.json.Json
+
 import scala.concurrent.Future
 
 object SendConfirmationEmail extends Logging {
@@ -24,15 +25,15 @@ object SendConfirmationEmail extends Logging {
     currency: Currency,
     directDebit: Option[DirectDebit],
     amountMinorUnits: Int
-  ): AsyncApiGatewayOp[ETPayload] = {
+  ): AsyncApiGatewayOp[ETPayload[ContributionFields]] = {
+    val maybeContributionFields = ContributionFields.fromData(amountMinorUnits, now, currency, directDebit, contacts)
 
-    ETPayload.fromData(amountMinorUnits, now, currency, directDebit, contacts) match {
-      case Some(payload) => AsyncApiGatewayOp(ContinueProcessing(payload))
-      case None =>
-        logger.info("skipping confirmation email, no email address in contact")
-        val response = ReturnWithResponse(ApiGatewayResponse.successfulExecution)
-        AsyncApiGatewayOp(response)
-
+    maybeContributionFields.map { fields =>
+      val payload = ETPayload(fields.EmailAddress, fields)
+      AsyncApiGatewayOp(ContinueProcessing(payload))
+    }.getOrElse {
+      logger.info("Not enough data in zuora account to send contribution thank you email, skipping")
+      AsyncApiGatewayOp(ReturnWithResponse(ApiGatewayResponse.successfulExecution))
     }
   }
